+++ es/components/LazyTextField.js
@@ -9,7 +9,7 @@ function _objectWithoutProperties(source, excluded) { if (source == null) return
 function _objectWithoutPropertiesLoose(source, excluded) { if (source == null) return {}; var target = {}; var sourceKeys = Object.keys(source); var key, i; for (i = 0; i < sourceKeys.length; i++) { key = sourceKeys[i]; if (excluded.indexOf(key) >= 0) continue; target[key] = source[key]; } return target; }
 
 import Input from "antd/lib/input";
-import { createComponent } from "./createComponent";
+import { createComponent } from "../createComponent";
 import { getValidateStatus } from "../maps/mapError"; // will trigger on change only onBlur
 // usefull for performance reasons
