+++ es/components/TextAreaField.js
@@ -1,5 +1,5 @@
 import Input from "antd/lib/input";
 import textFieldMap from "../maps/textFieldMap";
-import { createComponent } from "./createComponent";
+import { createComponent } from "../createComponent";
 var TextArea = Input.TextArea;
 export default createComponent(TextArea, textFieldMap);
