+++ es/components/TextField.js
@@ -1,4 +1,4 @@
 import Input from "antd/lib/input";
 import textFieldMap from "../maps/textFieldMap";
-import { createComponent } from "./createComponent";
+import { createComponent } from "../createComponent";
 export default createComponent(Input, textFieldMap);
