+++ lib/components/CheckboxField.js
@@ -7,7 +7,7 @@ exports["default"] = void 0;
 
 var _checkbox = _interopRequireDefault(require("antd/lib/checkbox"));
 
-var _createComponent = require("./createComponent");
+var _createComponent = require("../createComponent");
 
 var _eventMap = _interopRequireDefault(require("../maps/eventMap"));
