+++ lib/components/DatePicker.js
@@ -11,7 +11,7 @@ var _datePicker = _interopRequireDefault(require("antd/lib/date-picker"));
 
 var _mapError = require("../maps/mapError");
 
-var _createComponent = require("./createComponent");
+var _createComponent = require("../createComponent");
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
