+++ lib/components/LazyTextField.js
@@ -7,7 +7,7 @@ exports["default"] = void 0;
 
 var _input = _interopRequireDefault(require("antd/lib/input"));
 
-var _createComponent = require("./createComponent");
+var _createComponent = require("../createComponent");
 
 var _mapError = require("../maps/mapError");
