+++ lib/components/NumberField.js
@@ -9,7 +9,7 @@ var _inputNumber = _interopRequireDefault(require("antd/lib/input-number"));
 
 var _mapError = _interopRequireDefault(require("../maps/mapError"));
 
-var _createComponent = require("./createComponent");
+var _createComponent = require("../createComponent");
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
