+++ lib/components/RadioField.js
@@ -7,9 +7,9 @@ exports["default"] = void 0;
 
 var _eventMap = _interopRequireDefault(require("../maps/eventMap"));
 
-var _withOptions = require("./withOptions");
+var _withOptions = require("../withOptions");
 
-var _createComponent = require("./createComponent");
+var _createComponent = require("../createComponent");
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
