+++ lib/components/SliderField.js
@@ -9,7 +9,7 @@ var _slider = _interopRequireDefault(require("antd/lib/slider"));
 
 var _mapError = require("../maps/mapError");
 
-var _createComponent = require("./createComponent");
+var _createComponent = require("../createComponent");
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
