+++ lib/components/SwitchField.js
@@ -9,7 +9,7 @@ var _switch = _interopRequireDefault(require("antd/lib/switch"));
 
 var _mapError = require("../maps/mapError");
 
-var _createComponent = require("./createComponent");
+var _createComponent = require("../createComponent");
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
