+++ lib/components/TextField.js
@@ -9,7 +9,7 @@ var _input = _interopRequireDefault(require("antd/lib/input"));
 
 var _textFieldMap = _interopRequireDefault(require("../maps/textFieldMap"));
 
-var _createComponent = require("./createComponent");
+var _createComponent = require("../createComponent");
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
