+++ lib/index.js
@@ -94,7 +94,7 @@ Object.defineProperty(exports, "TextField", {
   }
 });
 
-var _createComponent = require("./components/createComponent");
+var _createComponent = require("./createComponent");
 
 var _mapError = require("./maps/mapError");
