+++ src/components/NumberField.js
@@ -1,5 +1,5 @@
 import InputNumber from "antd/lib/input-number";
 import mapError from "../maps/mapError";
-import { createComponent } from "./createComponent";
+import { createComponent } from "../createComponent";
 
 export default createComponent(InputNumber, mapError);
