+++ src/components/RadioField.js
@@ -1,5 +1,5 @@
 import eventMap from "../maps/eventMap";
-import { RadioField as Radio } from "./withOptions";
-import { createComponent } from "./createComponent";
+import { RadioField as Radio } from "../withOptions";
+import { createComponent } from "../createComponent";
 
 export default createComponent(Radio, eventMap);
