+++ src/createComponent.js
